feat (refs T31120): Move business logic to new service
